loreo2jps: Handle .jpg input
If we have already done some RAW processing on the image from the camera
(or for some reason aren't shooting in RAW) we may wish to convert the
JPEG instead of the RAW image, so handle JPEGs on the command line.

+++ photo-gallery/loreo2jps.sh
@@ -18,14 +18,28 @@
 overlap=5
 
 for file in "$@"; do
-	dcraw -e "$file"
 	base=$(basename "$file" .CR2)
-	jpg="${base}.thumb.jpg"
+	do_rm=0
+	if [ "$base" != "$file" ]; then
+		dcraw -e "$file"
+		jpg="${base}.thumb.jpg"
+		do_rm=1
+	else
+		base=$(basename "$file" .jpg)
+		if [ "$base" = "$file" ]; then
+			echo "Unknown file type"
+			continue
+		fi
+		jpg="${file}"
+	fi
 	jps="${base}.jps"
 	width=$(identify -format '%w' "$jpg")
 	height=$(identify -format '%h' "$jpg")
 	w2=$((($width / 2) + ($width * $overlap / 200)))
 	x2=$((($width / 2) - ($width * $overlap / 200)))
+	echo "$file -> $jps"
 	convert "$jpg[${w2}x${height}+${x2}+0]" "$jpg[${w2}x${height}+0+0]" +append "$jps"
-	rm "$jpg"
+	if [ "$do_rm" -eq 1 ]; then
+		rm "$jpg"
+	fi
 done
